+++ packages/shared-base/src/Site/index.ts
@@ -1,7 +1,7 @@
 import { getEnumAsArray } from '@masknet/kit'
 import { Sniffings } from '../Sniffings/index.js'
 import { ExtensionSite, EnhanceableSite } from './types.js'
-import { parseURL } from '../utils/parseURL.js'
+import { parseURL } from '../helpers/parseURL.js'
 
 const matchEnhanceableSiteHost: Record<EnhanceableSite, RegExp> = {
     [EnhanceableSite.Localhost]: /localhost/i,
