+++ packages/shared-base/src/Telemetry/index.ts
@@ -1,3 +1,3 @@
-import { ValueRef } from '../utils/ValueRef.js'
+import { ValueRef } from '../helpers/ValueRef.js'
 
 export const TelemetryID = new ValueRef('0000000000000000000000000000000000000000')
