+++ packages/shared-base/src/helpers/attachRectangle.ts
