+++ packages/shared-base/src/helpers/createLookupTableResolver.ts
