+++ packages/shared-base/src/helpers/getDefaultWalletName.ts
