+++ packages/shared-base/src/helpers/getDefaultWalletPassword.ts
