+++ packages/shared-base/src/helpers/index.ts
@@ -13,8 +13,3 @@ export * from './resolve.js'
 export * from './subscription.js'
 export * from './getDomainSystem.js'
 export * from './parseURL.js'
-export * from './generateContactAvatarColor.js'
-
-export type PartialRequired<T, RequiredKeys extends keyof T> = Omit<T, RequiredKeys> & Pick<Required<T>, RequiredKeys>
-
-export type UnboxPromise<T> = T extends Promise<infer U> ? U : never
