+++ packages/shared-base/src/index.ts
@@ -1,6 +1,9 @@
 export * from '@masknet/base'
 
-export * from './constant.js'
+export * from './constants.js'
+export * from './types.js'
+export * from './helpers/index.js'
+
 export * from './Account/index.js'
 export * from './MIME/index.js'
 export * from './Messages/Events.js'
@@ -29,9 +32,6 @@ export * from './kv-storage/index.js'
 export * from './onDemandWorker/index.js'
 export * from './SingletonModal/index.js'
 export * from './serializer/index.js'
-export * from './test-helpers/index.js'
-export * from './unsafe-html/index.js'
-export * from './utils/index.js'
 export * from './legacy-settings/index.js'
 export * from './settings/index.js'
 export * from './Telemetry/index.js'
