+++ packages/shared-base/src/legacy-settings/createSettings.ts
@@ -1,6 +1,6 @@
 import { type Option, None } from 'ts-results-es'
 import { MaskMessages } from '../Messages/index.js'
-import { type ValueRef, ValueRefJSON, ValueRefWithReady, type ValueComparer } from '../utils/index.js'
+import { type ValueRef, ValueRefJSON, ValueRefWithReady, type ValueComparer } from '../helpers/index.js'
 
 let getValue: (key: string) => Promise<Option<any>> = async () => {
     return None
