--- packages/shared-base/src/test-helpers/constants.ts
@@ -1,2 +1,0 @@
-export const UNIT_TEST_ADDRESS = '0x732b8e42455f79F3072fe18222A7E926588B4747'
-export const UNIT_TEST_ERC20_ADDRESS = '0xdAC17F958D2ee523a2206206994597C13D831ec7'
