--- packages/shared-base/src/test-helpers/index.ts
@@ -1,1 +1,0 @@
-export * from './constants.js'
