+++ packages/shared-base/src/types.ts
@@ -1,0 +1,3 @@
+export type PartialRequired<T, RequiredKeys extends keyof T> = Omit<T, RequiredKeys> & Pick<Required<T>, RequiredKeys>
+
+export type UnboxPromise<T> = T extends Promise<infer U> ? U : never
